fix: likeCount int → long 타입으로 변경

+++ src/main/java/com/sprint/mission/monew/domain/comment/dto/response/CommentResponse.java
@@ -10,7 +10,7 @@ public record CommentResponse(
     UUID userId,
     String userNickname,
     String content,
-    int likeCount,
+    long likeCount,
     boolean likedByMe,
     Instant createdAt
 ) {
